requirements.txt
requests==2.23.0
dnspython==1.16.0
